Fix error "dh key too small" with OpenSSL 1.1.1
With the introduction of openssl version 1.1.1 in Debian several packages have become buggy. This is due to changes in openssl that may come up during run time. 
This is caused by the SECLEVEL 2 setting the security level to 112 bit. This means that RSA and DHE keys need to be at least 2048 bit long. SHA-1 is no longer supported for signatures in certificates and you need at least SHA-256. Note that CAs have stopped issuing certificates that didn't meet those requirements in January 2015, and since January 2017 all valid CA certificates should meet those requirements. However there are certificates generated by private CAs or that are in a test suite that do not meet those requirements.